+++ packages/@overeng/megarepo/nix/build.nix
@@ -3,10 +3,19 @@
 #
 # TODO: Move shell completion generation into mkPnpmCli helper
 # so all CLIs get completions automatically.
-{ pkgs, src, gitRev ? "unknown", commitTs ? 0, dirty ? false }:
+{
+  pkgs,
+  depsPkgs ? pkgs,
+  src,
+  gitRev ? "unknown",
+  commitTs ? 0,
+  dirty ? false,
+}:
 
 let
-  mkPnpmCli = import ../../../../nix/workspace-tools/lib/mk-pnpm-cli.nix { inherit pkgs; };
+  mkPnpmCli = import ../../../../nix/workspace-tools/lib/mk-pnpm-cli.nix {
+    inherit pkgs depsPkgs;
+  };
   lockfileHash = "sha256-ZHb6oroLpfvQzk7Cs3TuqX7GJR+34qzXD0pLmew3da4=";
   packageJsonDepsHash = "sha256-6FpelxhQZoLp6Gg5WvZAqjB4V4Zm8XQx0vFIEA9PtN8=";
   base = mkPnpmCli {